Saving daily habits entries

+++ ViewModels/HabitEntryVM.cs
@@ -5,5 +5,5 @@ public class HabitEntryVM
     public int HabitId { get; set; }
     public DateTime Date { get; set; }
     public bool Completed { get; set; }
-    public int? Quantity { get; set; }
+    public float? Quantity { get; set; }
 }
